#7397 - Refactoring
