+++ sormas-backend/src/main/java/de/symeda/sormas/backend/caze/CaseService.java
@@ -117,7 +117,6 @@ import de.symeda.sormas.backend.externaljournal.ExternalJournalService;
 import de.symeda.sormas.backend.feature.FeatureConfigurationFacadeEjb.FeatureConfigurationFacadeEjbLocal;
 import de.symeda.sormas.backend.hospitalization.Hospitalization;
 import de.symeda.sormas.backend.immunization.ImmunizationService;
-import de.symeda.sormas.backend.immunization.entity.Immunization;
 import de.symeda.sormas.backend.infrastructure.community.Community;
 import de.symeda.sormas.backend.infrastructure.district.District;
 import de.symeda.sormas.backend.infrastructure.facility.Facility;
@@ -148,7 +147,6 @@ import de.symeda.sormas.backend.util.IterableHelper;
 import de.symeda.sormas.backend.util.JurisdictionHelper;
 import de.symeda.sormas.backend.util.ModelConstants;
 import de.symeda.sormas.backend.util.QueryHelper;
-import de.symeda.sormas.backend.vaccination.VaccinationService;
 import de.symeda.sormas.backend.visit.Visit;
 import de.symeda.sormas.backend.visit.VisitFacadeEjb;
 import de.symeda.sormas.utils.CaseJoins;
@@ -181,8 +179,6 @@ public class CaseService extends AbstractCoreAdoService<Case> {
 	private TravelEntryService travelEntryService;
 	@EJB
 	private ImmunizationService immunizationService;
-	@EJB
-	private VaccinationService vaccinationService;
 	@EJB
 	private FeatureConfigurationFacadeEjbLocal featureConfigurationFacade;
 	@EJB
@@ -1779,20 +1775,6 @@ public class CaseService extends AbstractCoreAdoService<Case> {
 		cu.where(CriteriaBuilderHelper.and(cb, cb.equal(root.get(Case.PERSON), personId), cb.equal(root.get(Case.DISEASE), disease), datePredicate));
 
 		em.createQuery(cu).executeUpdate();
-	}
-
-	public void updateVaccinationStatuses(Case caze) {
-		List<Immunization> casePersonImmunizations = immunizationService.getByPersonAndDisease(caze.getPerson().getUuid(), caze.getDisease(), true);
-
-		boolean hasValidVaccinations = casePersonImmunizations.stream()
-			.anyMatch(
-				immunization -> immunization.getVaccinations()
-					.stream()
-					.anyMatch(vaccination -> vaccinationService.isVaccinationRelevant(caze, vaccination)));
-
-		if (hasValidVaccinations) {
-			caze.setVaccinationStatus(VaccinationStatus.VACCINATED);
-		}
 	}
 
 	private float calculateCompleteness(Case caze) {
