+++ sormas-backend/src/main/java/de/symeda/sormas/backend/vaccination/VaccinationFacadeEjb.java
@@ -30,6 +30,7 @@ import javax.validation.constraints.NotNull;
 
 import de.symeda.sormas.api.Disease;
 import de.symeda.sormas.api.caze.CaseReferenceDto;
+import de.symeda.sormas.api.caze.VaccinationStatus;
 import de.symeda.sormas.api.contact.ContactReferenceDto;
 import de.symeda.sormas.api.event.EventParticipantReferenceDto;
 import de.symeda.sormas.api.i18n.I18nProperties;
@@ -54,6 +55,7 @@ import de.symeda.sormas.backend.caze.CaseService;
 import de.symeda.sormas.backend.clinicalcourse.ClinicalCourseFacadeEjb;
 import de.symeda.sormas.backend.contact.Contact;
 import de.symeda.sormas.backend.contact.ContactService;
+import de.symeda.sormas.backend.event.Event;
 import de.symeda.sormas.backend.event.EventParticipant;
 import de.symeda.sormas.backend.event.EventParticipantService;
 import de.symeda.sormas.backend.immunization.ImmunizationEntityHelper;
@@ -358,6 +360,54 @@ public class VaccinationFacadeEjb implements VaccinationFacade {
 			caseService.updateVaccinationStatuses(personId, disease, newVaccinationDate);
 			contactService.updateVaccinationStatuses(personId, disease, newVaccinationDate);
 			eventParticipantService.updateVaccinationStatuses(personId, disease, newVaccinationDate);
+		}
+	}
+
+	public void updateVaccinationStatuses(Case caze) {
+		List<Immunization> casePersonImmunizations = immunizationService.getByPersonAndDisease(caze.getPerson().getUuid(), caze.getDisease(), true);
+
+		boolean hasValidVaccinations = casePersonImmunizations.stream()
+			.anyMatch(
+				immunization -> immunization.getVaccinations()
+					.stream()
+					.anyMatch(vaccination -> vaccinationService.isVaccinationRelevant(caze, vaccination)));
+
+		if (hasValidVaccinations) {
+			caze.setVaccinationStatus(VaccinationStatus.VACCINATED);
+		}
+	}
+
+	public void updateVaccinationStatuses(Contact contact) {
+		List<Immunization> contactPersonImmunizations =
+			immunizationService.getByPersonAndDisease(contact.getPerson().getUuid(), contact.getDisease(), true);
+
+		boolean hasValidVaccinations = contactPersonImmunizations.stream()
+			.anyMatch(
+				immunization -> immunization.getVaccinations()
+					.stream()
+					.anyMatch(vaccination -> vaccinationService.isVaccinationRelevant(contact, vaccination)));
+
+		if (hasValidVaccinations) {
+			contact.setVaccinationStatus(VaccinationStatus.VACCINATED);
+		}
+	}
+
+	public void updateVaccinationStatuses(EventParticipant eventParticipant) {
+		if (eventParticipant.getEvent().getDisease() == null) {
+			return;
+		}
+		List<Immunization> eventParticipantImmunizations =
+			immunizationService.getByPersonAndDisease(eventParticipant.getPerson().getUuid(), eventParticipant.getEvent().getDisease(), true);
+		Event event = eventParticipant.getEvent();
+
+		boolean hasValidVaccinations = eventParticipantImmunizations.stream()
+			.anyMatch(
+				immunization -> immunization.getVaccinations()
+					.stream()
+					.anyMatch(vaccination -> vaccinationService.isVaccinationRelevant(event, vaccination)));
+
+		if (hasValidVaccinations) {
+			eventParticipant.setVaccinationStatus(VaccinationStatus.VACCINATED);
 		}
 	}
